Moving and splitting pckage
