+++ biopet-tools/src/test/scala/nl/lumc/sasc/biopet/tools/VcfStatsTest.scala
@@ -15,11 +15,12 @@
 package nl.lumc.sasc.biopet.tools
 
 import java.io.File
-import java.nio.file.{ Files, Paths }
+import java.nio.file.{Files, Paths}
 
 import htsjdk.variant.variantcontext.Allele
 import htsjdk.variant.vcf.VCFFileReader
-import nl.lumc.sasc.biopet.tools.VcfStats._
+import nl.lumc.sasc.biopet.tools.vcfstats.VcfStats
+import nl.lumc.sasc.biopet.tools.vcfstats.VcfStats._
 import org.scalatest.Matchers
 import org.scalatest.testng.TestNGSuite
 import org.testng.annotations.Test
